Create 1_easy_twosum.java
Leetcode 1 question

Two sum question that i solved using brute force firstly then i started to look for other methods and found hashmap is the optimal solution for this question. I also tried two pointer method for the question but this method wont for because it sorts the array first.