Create 2_medium_addtwonumbers.java
Leetcode question 2

Two solutions for the question second one is my first try at the question and it has not been accepted as it is limited to long or int max value.
First solution attempted after failing. On this solution i have tried to sum the numbers digit by digit. This solution works without limit and therefore it has been accepted.